# backend/services/order/api/docs/swagger.yaml
definitions:
...
  types.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/order/list:
    get:
      consumes:
      - application/json
      description: 執行 ListHandler 動作
      parameters:
      - default: 1
        description: 頁碼
        in: query
        name: page
        type: integer
      - default: 20
        description: 每頁筆數
        in: query
        name: pageSize
        type: integer
      - description: 關鍵字
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      summary: ListHandler
      tags:
      - Order
...
swagger: "2.0"